+++ app/Http/Controllers/Authenticate/VilleController.php
@@ -6,14 +6,15 @@ use App\Http\Controllers\Controller;
 use App\Models\Region;
 use App\Models\Town;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Validator;
 
 class VilleController extends Controller
 {
 
     public function index(Request $request)
     {
-
+        if (!auth()->user()->is_admin) {
+            return back();
+        }
         $towns = Town::all();
         $regions = Region::all();
         return view('admin.authentication.layouts.pages.town.show', compact('towns', 'regions'));
@@ -22,6 +23,9 @@ class VilleController extends Controller
     public function store(Request $request)
     {
 
+        if (!auth()->user()->is_admin) {
+            return back();
+        }
         $request->validate([
             'name' => 'required|string|max:255'
         ]);
@@ -32,6 +36,9 @@ class VilleController extends Controller
 
     public function delete(Town $town)
     {
+        if (!auth()->user()->is_admin) {
+            return back();
+        }
         $town->delete();
 
         return redirect()->route('admin.town.index')->with('success', 'La ville a été supprimée avec succès.');
